+++ src/graphics/sprite.rs
@@ -204,37 +204,37 @@ impl<'s> Drawable for Sprite<'s> {
 
 impl<'s> Transformable for Sprite<'s> {
     fn set_position<P: Into<Vector2f>>(&mut self, position: P) {
-        unsafe { ffi::sfSprite_setPosition(self.sprite.as_ptr(), position.into().raw()) }
+        unsafe { ffi::sfSprite_setPosition(self.sprite.as_ptr(), position.into()) }
     }
     fn set_rotation(&mut self, angle: f32) {
         unsafe { ffi::sfSprite_setRotation(self.sprite.as_ptr(), angle) }
     }
     fn set_scale<S: Into<Vector2f>>(&mut self, scale: S) {
-        unsafe { ffi::sfSprite_setScale(self.sprite.as_ptr(), scale.into().raw()) }
+        unsafe { ffi::sfSprite_setScale(self.sprite.as_ptr(), scale.into()) }
     }
     fn set_origin<O: Into<Vector2f>>(&mut self, origin: O) {
-        unsafe { ffi::sfSprite_setOrigin(self.sprite.as_ptr(), origin.into().raw()) }
+        unsafe { ffi::sfSprite_setOrigin(self.sprite.as_ptr(), origin.into()) }
     }
     fn position(&self) -> Vector2f {
-        unsafe { Vector2f::from_raw(ffi::sfSprite_getPosition(self.sprite.as_ptr())) }
+        unsafe { ffi::sfSprite_getPosition(self.sprite.as_ptr()) }
     }
     fn rotation(&self) -> f32 {
         unsafe { ffi::sfSprite_getRotation(self.sprite.as_ptr()) }
     }
     fn get_scale(&self) -> Vector2f {
-        unsafe { Vector2f::from_raw(ffi::sfSprite_getScale(self.sprite.as_ptr())) }
+        unsafe { ffi::sfSprite_getScale(self.sprite.as_ptr()) }
     }
     fn origin(&self) -> Vector2f {
-        unsafe { Vector2f::from_raw(ffi::sfSprite_getOrigin(self.sprite.as_ptr())) }
+        unsafe { ffi::sfSprite_getOrigin(self.sprite.as_ptr()) }
     }
     fn move_<O: Into<Vector2f>>(&mut self, offset: O) {
-        unsafe { ffi::sfSprite_move(self.sprite.as_ptr(), offset.into().raw()) }
+        unsafe { ffi::sfSprite_move(self.sprite.as_ptr(), offset.into()) }
     }
     fn rotate(&mut self, angle: f32) {
         unsafe { ffi::sfSprite_rotate(self.sprite.as_ptr(), angle) }
     }
     fn scale<F: Into<Vector2f>>(&mut self, factors: F) {
-        unsafe { ffi::sfSprite_scale(self.sprite.as_ptr(), factors.into().raw()) }
+        unsafe { ffi::sfSprite_scale(self.sprite.as_ptr(), factors.into()) }
     }
     fn transform(&self) -> &Transform {
         unsafe { &*ffi::sfSprite_getTransform(self.sprite.as_ptr()) }